Extract InboxThread and userType to seperate files
