+++ modules/users/client/users.prop-types.js
@@ -1,0 +1,7 @@
+import PropTypes from 'prop-types';
+
+export const userType = PropTypes.shape({
+  _id: PropTypes.string.isRequired,
+  username: PropTypes.string.isRequired,
+  displayName: PropTypes.string.isRequired,
+});
